bugfix: fixed problems with type convertion

+++ src/vm_setup/linux_setup.rs
@@ -6,7 +6,7 @@
 use kvm_ioctools::{Kvm, VcpuExit, Vcpu};
 use crate::vm_setup::setup_utils::VmSetup;
 use vm_memory::{GuestAddress, GuestMemoryMmap};
-use kvm_bindings::{kvm_userspace_memory_region};
+use kvm_bindings;
 
 /// Asynchronously runs a virtual machine using KVM with the provided setup.
 ///
@@ -44,7 +44,7 @@ pub async fn run_vm(setup: VmSetup) -> Result<(), String> {
     if let Err(e) = vm.set_memory_region(kvm_bindings::kvm_userspace_memory_region {
         slot: 0,
         guest_phys_addr: 0x4000,
-        memory_size: setup.get_memory_size(),
+        memory_size: setup.get_memory_size() as u64,
         userspace_addr: host_addr,
         flags: 0,
     }) {
@@ -52,7 +52,8 @@ pub async fn run_vm(setup: VmSetup) -> Result<(), String> {
     };
 
     // Spawn a blocking task for each virtual CPU core
-    let handlers: Vec<tokio::task::JoinHandle<Result<String, String>>> = Vec::with_capacity(setup.get_cpu_cores_count());
+    let handlers: Vec<tokio::task::JoinHandle<Result<String, String>>> =
+        Vec::with_capacity(setup.get_cpu_cores_count() as usize);
     for cpu_id in 0..setup.get_cpu_cores_count() {
         // Create a VCPU for this core
         let vcpu = match vm.create_vcpu(cpu_id) {
@@ -68,7 +69,7 @@ pub async fn run_vm(setup: VmSetup) -> Result<(), String> {
 
         regs.rip = 0x1000;
         regs.rflags = 0x2;
-        
+
         if let Err(e) = vcpu.set_regs(&regs) {
             return Err(format!("Failed to set VCPU {} registers: {}", cpu_id, e));
         };
